Move icons, replace favicon icon
