+++ src/10_styles/_design-tokens.scss
@@ -59,7 +59,7 @@
 // 005 Iconography
 /*******************/
 // Icon Design Tokens
-@import "../00_design_tokens/005_iconography/Icon/Icon";
+@import "../00_design_tokens/005_iconography/Icons/Icons";
 
 // 006 Typography
 /******************/
