+++ src/10_styles/main.scss
@@ -17,6 +17,9 @@
 // Global (Root and Body)
 @import "../10_styles/global";
 
+// Iconography
+@import "../00_iconography/Icon";
+
 // Typography Styles
 @import "../00_typography/Typography";
